Clean up whitespace in settings files
Whitespace should be consistent inside of local_settings.py.example.

Currently, most comments in side our settings files follow the "space
after # mark" guideline. This is as it should be. However, commented out
code-lines are inconsistently including that space. Most of the time
they don't include it.

There's a case to be made for either approach:

Do include the space after the #:
    Keeps things consistent and it's how pyflakes likes things.

Do not include the space after the #:
    Identifies the commented out line as potential code, instead of a
    descriptive comment.

    Makes it easier to just remove the '#' character to activate the
    line of code. Specifically with python, leaving the white space is
    likely bad.

An informal survey of the lines in these files shows that most commented
out code lines do not have a space, and I agree with that policy. I'd
like to make this the policy and clean up the code.

Also added trailing ',' chars on the last element of many lists.

Change-Id: I36b6b40996f7ee5eb566719fe7a92ddecf569f5d
Closes-Bug: #1428222