Fix tutorial to correctly display custom dashboard/panel
This change fixes a bug in the "Building a Dashboard using
Horizon" tutorial in the Horizon developer documentation.
The tutorial shows new Horizon developers how to build a
custom dashboard and panel that displays running instances.

In the tutorial, the URL pattern provided fails to display
the custom dashboard & panel. This change updates the URL
pattern. The custom dashboard and panel now correctly show
the custom view of instances data.

Documentation has been regenerated and reviewed to ensure
change was correctly done.

DocImpact
Closes-Bug: #1427758

Change-Id: I522825c6d24fcaf0981f5faab273e2efd6e00f1a