Floating IP table should support sorting
Currently the IPv4Address parser in the tablesort only look up the text inside
<li></li>. This works fine in the instance table since the cell format has a list to
wrap around it. However, other table can still fit the IP without <li> with it.

This patach is to enhance the parser to look up the text in the cell directly if <li>
is not presented.

Change-Id: Id22ebc538d6d2adbcb6be21be2f28c47b0fcc217
Closes-Bug: #1428172