[Launch Instance Fix] Source Details - Metadata
The metadata display widget was developed as a common way to
display additional metadata about images, volumes, and flavors
in the launch instance workflow. This fix integrates the metadata
widget into the source step for displaying the extra metadata.

It also corrects any additional styling problems in the widget itself.

The bug report includes detailed information on testing.

Closes-Bug: 1433792
Co-Authored-By: Szymon Wroblewski <szymon.wroblewski@intel.com>
Co-Authored-By: Brian Tully <brian.tully@hp.com>
Change-Id: I547ccd6ccd65116fc8ac2e77d438bbf8c37a7afd