Changing terminology for jobs and job executions in data processing
The data processing service is now using the term "Job
Template" for the entity that was formerly called "Job".
Similarly, what was formerly known as  a "Job Execution
is now known as a "Job".  These terms are in-line with
the CLI.  The underlying code still references the old
names since the api remains unchanged.

Change-Id: I47975e4ef54268fa5500328939c66577659b6deb
Partial-Implements: blueprint data-processing-rework-ui