Source step - fix instance count limit and donut chart
In the current version of the Launch Instance Wizard Source Step,
the instance count field is not correctly limited based on current
instance usage and nova limits (maxTotalInstances). There is a bug
whereby the user can enter an instance count of more than what is
available (exceeding their quota). In addition, the donut chart
does not update properly when this occurs. This patch corrects this
so that if a user enters a number greater than what is available,
it automatically changes the count to the maximum count that will still
stay within quota limits.

Closes-Bug: 1432897

Change-Id: I24f1309b04f3c045c02b67409c49fc3ee86aea3f