Metadata widget support for readonly flag
Metadata widget doesn't handle the `readonly` flag for properties which is
returned by Glance API. If the property is defined as `readonly` the input
should stick to the default value of the property.

The readonly flag has been introduced in Glance in bug report:
https://bugs.launchpad.net/glance/+bug/1406235

This patch adds support for readonly flag in metadata widget.

Closes-Bug: 1433026
Change-Id: I79e2b7c8a397f6b19659d6e13f7fad29bc26892e