[Launch Instance Fix] Add details to Select Flavor step
Add details to the Select Flavor step of the Angular-based launch
instance wizard. Details are charts indicating how much each flavor
affects the current user quota.

This commit also refactors the flavor step to not duplicate HTML
used for the allocated and available flavor tables. Instead a
directive provides a"flavor table" which is then reused for both
the allocated and available sections the parent transfer table.

Change-Id: I446051463d83aba9ecb0d862ca6c5b1e1b2181a9
Partial-Bug: 1433373