[Launch Instance Fix] Enhancing wizard widget
1) Add some events to Wizard widget to provide hooks for logic in model(s)
and controller(s) for each step.

New wizard events added are:
- ON_INIT_SUCCESS
- ON_INIT_ERROR
- ON_SWITCH
- BEFORE_SUBMIT
- AFTER_SUBMIT

Some features in Launch Instance work flow can take advantage of this
infrastructure change:

- Refreshing page on `AFTER_SUBMIT` event.
- Lazy data sync and validation on `BEFORE_SUBMIT` event.
- Dynamical showing/hiding optional steps/content according to async check for
  availability of certain services on `ON_INIT_SUCCESS` event.

2) Add a new feature to check readiness for each step that provides a optional `checkReadiness`
method. Each such a method should return a promise. When this method is
provided with the step, the `ready` property of the step will be set to
`false` until the promise gets resolved. If no `checkReadiness` method
provided, the `ready` property of the step will be set to `true` directly.

This it useful for those work flow where some step may be optional and
should be showing to UI conditionally and the check for the condition
is an asynchronous operation.

Co-Authored-By: Brian Tully <brian.tully@hp.com>
Closes-Bug: 1433674
Change-Id: I556e263f4c701a378452b4397585f1b4d8abb763