Global error HTTP handler
Currently in angular, we are using ajax to make various REST api calls. When
a user session expires, there is no way to know this.  Consequently, when
the user attempt to do something a 401 is returned.

Instead of checking for this 401 in all of our REST invocation, it is best to do
a global error check for 401. If this happen, we log the user out and redirect
the user to the login screen.

In addition, policy alerts will now auto dismiss.

Partially Implemenents: blueprint angularize-identity-tables

Change-Id: I427977b9cf7eeee928e1ade9355954bbb385170c