Modified hzSelectAll to watch changes in row length
hzSelectAll run once initially and does not know about new rows that are
added. For tables that require adding or removing of rows, this poses a
significant problem.

To solve this issue, we need to watch the changes made to row. However,
watching the entire collection is also very costly. Instead, we will watch
over the change in row length.

Partially Implemenents: blueprint angularize-identity-tables
Change-Id: I2a6718737b6855181f483fc63d42637a0cec548d