Removing unnecessary and annoying error message
In network topology, a server in the error state has no console
available to connect to. The page refreshes via an ajax request
every 10 seconds. When it does, a new error message gets added to
the screen each time. The list quickly becomes very long. The
message is also not over helpful "No available console found."
Logging is not useful either, so just ignoring error.

Closes-Bug: #1436627
Change-Id: I1490ef7f6157b17113692dd6baa4a6f91e01fe8a