fixed row update of vpn table
In vpn table, after creating a vpn the row should be updated
whenever the status of the vpn changes. but it is not and it get
stuck in the pending_create status.

It needs page reload to update the row.

In this patch, the added Update row class will reload
a row after the status change.

Change-Id: I45e3e52e72ac3cadbdcf58ea540281ba126d9b49
Closes-Bug: #1438088