Fix date pickers in metering modal
The metering forms were using the date picker functionality from the
d3 line chart library, this used to work when it was loaded as a whole
page as the d3 libary would run intending to setup any d3 graphs however
now it is being loaded as a modal the elements that would be affected by
the d3 code do not exist at page load time.

In order to get the date picker events to be added to the elements on
the form when the modal loads, the JS in the forms template has been
extended in this patch to support adding those events.

This patch also changes the data attributes on the form elements so that
they do not conflict with any d3 line chart elements.

Alongside the javascript this patch fixes bug to do with parsing the
dates in metering utils and adds UTs to prevent regression.

Change-Id: I4e239daa03b2f54e434254bac48ba0cceb037b5d
Closes-Bug: 1427756