[Launch Instance Fix] Flavor Metadata Details
The metadata display widget was developed as a common way to
display additional metadata about images, volumes, and flavors
in the launch instance workflow. This fix integrates the metadata
widget into the source step for displaying the extra metadata.

It also corrects any additional styling problems in the widget itself.

The bug report includes detailed information on testing.

Closes-Bug: #1438505
Co-Authored-By: Szymon Wroblewski <szymon.wroblewski@intel.com>
Change-Id: Id84548cfaaa50a3d1e299b34b3a2591ec02ed25b