added common name based filter for tables
In horizon most number of the tables have the name based client side filter.
All the tables are implementing the filter class individually,
but the logic and code is same in all the places.
Adding generic class for name based filter will remove lot of the duplicate code.

In this patch i have added the generic class and using it in all the vpn tables,
once it is accepted, we can remove the duplicate classes
and use the generic class everywhere.

Change-Id: Ie9c6bfaaffc72f8144fa36d450cce77b16fae64c
Implements: blueprint common-client-filter
Closes-Bug: #1427521