Remove concatenation in csv file
The download CSV summary file from both the Admin and Project
overview pages includes concatenation in its contents.
The : needs to be included as part of the label so translators
have control of spacing. Note that the commas are part of the
CSV file format and should not be translated.

Change-Id: I13137fda9b3e93edf97a7539044e791d04017ea7
Closes-Bug: 1444135