Bind Launch instance quota chart to count input
For the Launch Instance source step, this patch binds the quota chart
directly to the instance count form input vs. only updating after a
source has been allocated. Since the selected source type does not
affect quota, it was a confusing experience when increasing the
instance count did not update the quota chart.

Change-Id: Ia99a61dd3a0769555850599f283b4fd2d401dc9c
Closes-Bug: 1443622