corrected the create volume from snapshot
While creating the volume from snapshot via horizon UI, the volume
is not created from snapshot. It is just creating a volume with
same size of the snapshot.

Fixed the condition used if snapshot_id should be passed when
creating volume. The volume_source_type field is empty string if
launched from Volume Snapshot panel.

Co-Authored-By: Lin Hua Cheng <os.lcheng@gmail.com>

Change-Id: I19ddd572d1a7a028e3ae4dae412ffc54e730e50c
Closes-Bug: #1447288