Compatibility fix for pyscss 1.3.4
This commit alters the action-list.scss file so it is compatible
with the newer version of pyScss (1.3.4+), allowing us to
upgrade to that version.

This change is a prerequisite for the update to get into the
global requirements so that that change can pass testing.

Change-Id: Idec7ce039b57476ce20b7b678431459c2ca0bdbf
Co-Authored-By: Kelly Domico <kelly.domico@hp.com>
(cherry picked from commit cca93ade7c23c1f2794376161b6660a459292eee)