ngReorg - Align constants, factories and services
This commit makes injectable names consistent to include
the module as a prefix. This reduces confusiong about
exactly which item is being injected.

It also simplifies the naming scheme for the Angular-based
API modules.

This is one step in a larger effort to restructure the Angular
source. See https://review.openstack.org/#/c/176152/ for the
full set of planned changes.

Change-Id: I87d4ff78b97950248442056ce3ce5af9177042ae
Partial-Bug: #1454880