ngReorg - move dashboard-app to dashboard
Move the angular app module from the framework to
openstack_dashboard at the top level as "horizon.app".

This edit also removes the old hzUtils.log function: it
exists in the framework, but is configured in the application,
thus causing dependency issues. It is only used in one place
(hz.tables.js), and for debugging. I contend that if debugging
is needed then $log can be asked for as needed.

The hzConfig.debug flag was only used by that one function,
so it is also removed in this edit.

Change-Id: Ie538940cd8d1b0eabe677790bad979cc146bfbd1
Closes-Bug: #1458697
Closes-Bug: #1465885