User should be able to update allocation pool values.
Steps to reproduce it:
1. Select an existing subnet.
2. In the subnet detail, change the allocation pools value.
3. Click Save.
4. Click the changed subnet detail and check the allocation pool value.

Root cause: The problem is that we only allow a new subnet to set the allocation pool.
Seems like we can do that via api during update and so we should enable that on Horizon.

Change-Id: I9e6b0ee427ca9af82c4ed80abc1e7c19fe81641f
Closes-bug: 1458803
Closes-bug: 1458769
(cherry picked from commit 3b6dacca71e0cfb69786a7b5713a151f59d2682d)