Add support for Docker image format
The 'Docker' format is added to the default list of supported
formats so it can be selected from the list of formats when creating
or updating an image. When the docker format is selected, the
disk_format and container_format properties are set appropriately.
This also includes changes to the Format column and filter for the
Images table to handle this type.

Closes-Bug: #1455179
Change-Id: Ie705c86d47c0b6035373b9311454748122185988