Adding download for openrc file for keystone v3
The existing openrc file download only works for keystone v2. Regardless
of whether v3 is enabled in Horizon.

This adds support for both. A v2.0 and a v3 compatible openrc file
download. A couple of different situations are covered.

1) support for keystone v2 only: OPENSTACK_API_VERSION={'identity': 2.0}
In this case only the v2 option is shown.

2) Use of keystone v3 in a potentially mixed environment. Since it is
possible to use keystone v2 and v3 in the same enviroment, having
OPENSTACK_API_VERSION={'identity': 3} displays options for downloading
v2 or v3 compatible openrc files.

Rationale for making the existing methods and urls support v3+. By
moving the v2.0 functionality to new version specific methods, they can
be more easily excised when v2 is obsolete and we're left with the newer
version support.

Change-Id: I29c62dc7436cc39adc1a4af9d90ceb6767e7a177
Closes-Bug: #1460150