Prevent multiple form submission on table row actions
This disables any table row actions that do a form submit after an
action is taken and the form is being submitted. It also changes the
cursor to make it clear an action is taking place.

Closes-Bug: #1433849
Change-Id: I004f0ae43e5201eba30ff38a34be680416407fbc
(cherry picked from commit ea2f6b35da70d7a9d3f441553dac0f88f6ef9218)

Handle modals properly when disabling table actions

This fixes a problem when attempting to disable table actions on
a modal dialog. It causes the Launch Instance modal to render
incorrectly and other modals result in a
"$form.on is not a function" error displayed in the browser
console. This is because the modal init function is assuming the
argument passed in is a jQuery reference to the modal but in
reality it is just the modal DOM node itself. The problem is fixed
by making a jQuery reference to the modal before preparing to
disable the actions of any tables that might be displayed on it.

Closes-Bug: #1462544
Change-Id: I56399468ff5e2b5bcd0b285ece0d6dd001500c24
(cherry picked from commit 54ee33c4d844961f7f7f7d94b3c298a015b09e95)