Handle errors during glance image updates correctly
Before an UnboundLocalError would occur in case an error would happen
during image updates, because of incorrect referencing of a variable.
Moreover the error was muted by exceptions.handle with ignore=True,
and could result in an incorrect success message.
This change removes references to unused image variable and allows
exceptions to be handled by the function, that calls image_update.

Change-Id: I35ea98a146690565582dd9beea62941f97d32474
Closes-Bug: #1464991