Add selective ESLint rule activation for Horizon Legacy JS
This patch adds the files for eslint overrides in the two different
segments of horizon's codebase: Legacy and Dashboard. There are
quite a few rules that had to be deactivated, which will reduce
the effectiveness of legacy linting.

Note: You cannot disable plugins in .eslintrc files, you can
only enable them. This is why there appear to be two redundant
.eslintrc files that enable the angular rules.

Change-Id: I5768cbc43e788a4320a6bb637c263dc3c3df3db9
Partially-Implements: blueprint jscs-cleanup