Theming Mechanism now supports 3rd party themes
In order to facilitate the bulk of the CSS Reorg effort, it was
extremely beneficial to quickly toggle between many different themes
in order to validate proper cascading inheritance.  This work
prompted the following example theme.

The 'material' theme gives an example of how to make use of a 3rd
party theme using the theming functionality.  In addition to
incorporating a theme, loaded as a static asset using
requirements.txt, it also gives examples of how to cleanly override
styles, variables, icon fonts and Django templates.

This theme is replacing 'blue' as the example of how to use a theme
other than 'default'.

To use this theme, add the theme to your local_settings.py:
CUSTOM_THEME_PATH = 'themes/material', recollect and recompress your
static files.

Change-Id: Ic67189de5aac5ca541ad6fe82b823e8fcf318bd0
Partially-Implements: blueprint horizon-theme-css-reorg