Add Angular Basic Page Header Directive
This provide an angular basic page header directive
and adds it to the ng images page.

We're addressing future tech debt here by breaking out the predicted
common pattern of page header HTML into a single angular directive.

This allows for templating with parameterization, so that if in the future
we or a deployer would like to customize all of their page headers,
they only need to do it in the angular template HTML.  It also makes
the usage of it very declarative in nature.

This is basically an angular based way to replace the current django template:
horizon/templates/horizon/common/_page_header.html

To test set DISABLED = False in _203_project_images_panel.py

Then go to the last images panel in the compute menu.

Change-Id: Ia753905f805af72535bd255887494a6954546b56
Partially-Implements: blueprint angularize-images-table