Adding Angular Project Images Table
This patch adds the images table to the projects panel.

It follows the pattern established by the following review:
https://review.openstack.org/#/c/197373/

However, it pre-fixes the table files with images- to make them
easier to find via search in tools like the pycharms debugger
and it the javascript console on various browsers.

To be added in subsequent patches:
 - Actions
 - Forms
 - Common images table directive (TBD)
 - Filters via Magic Search
 - Integration to Searchlight

To test set DISABLED = False in _203_project_images_panel.py

Change-Id: I5e509fb807c356d3fb4b58e974ffa160828dbf88
Partially-Implements: blueprint angularize-images-table