Improve error message for creating QoS Spec
When input a "QoS Spec" name that already exists in "create
QoS Spec" modal form, we will see the error message:
"Unable to create QoS Spec" in the top right corner of the
page, this error message is not specific and clear. If we
provide specific and clear error message: 'QoS Spec name "%s"
already exists.' beside the name field, it will be more better.

Change-Id: I3aed7b39fb7f3e991738ad246585af68314a0cf3
Closes-Bug: #1494166