Fix static file paths in heat topology view
The svg and gif images are still using '/static/' as the base url. Since
WEBROOT is configurable and STATIC_URL is as well. This is needs to be
fixed or the images won't be found when either has been changed from the
default.

Not a clean cherry-pick as additional image files were added in Liberty.
This just updates the URLs for those present in Kilo.

Conflicts:
        openstack_dashboard/dashboards/project/stacks/mappings.py

Closes-Bug: #1474512
Change-Id: I532216c86e131b947e4377303ff3123f32e236f9
(cherry picked from commit fce980c8ce62a66800d6afb7b2864cb4a99b4672)