Removing Network Profile support in stable/kilo
Network profile is not supported in stable/kilo by the Cisco N1KV
ML2 driver. As the patch for enabling/disabling network profile
support based on a configurable parameter was rejected, this patch
pulls out the network profile code completly from stable/kilo.

This patch is necessary for horizon to be functional with Cisco
N1KV ML2 driver.

Change-Id: I1ef2f489c9bb2adfbe3174c3667539c59562966f
Closes-bug: #1475190