Support create subnet w/Neutron subnet allocation
The user is enabled to choose between entering a network address
manually or to allocate a network address from a pool. With the
advent of address scopes in Neutron this feature will be used for
enabling routing of globally unique addresses (GUA) all the way in
to the VM. This is a prerequisite for IPv6 to be routable, and will
also at some point allow GUA-IPv4 without NAT.

Workflow:
If subnet allocation is not supported in backend, there are no
subnet pools available, or the user chooses to enter network address
manually the form will behave like before.

If subnet allocation is supported in backend and there either are
subnet pools available or the label for the default subnet pool(s) is
configured, the user may choose to allocate a network address from a
pool. The user will be presented with two new drop-down menus. The
first drop-down menu will, in addition to the default pool (if
configured), present the names and prefixes of available pools. The
next drop-down menu will display available choices for prefix lengths
(it will not be displayed if the "default pool" is selected). The
available choices are defined by the subnet pool. Default prefix length
defined by the pool will be selected. The correct IP version is used
based on what IP version is defined by the selected subnet pool.

The "Create Subnet" checkbox has been moved from the Subnet step to
the first step of the workflow. This is because of how switchable
fields work, and how you can not safely define more than one slug in a
fields "switch-on" attribute. For the sake of User Experience, we need
to reliably switch fields based on selected Address source on the
Subnet tab. IMHO it also makes more sense to have this checkbox on the
first step of the workflow, if you just want to create a network you can
now uncheck the "Create Subnet" checkbox and click directly on "Create".

Change-Id: Ie392ccd9feae4dc02c3f30e2475457e755700b6b
Implements: blueprint neutron-subnet-allocation