Dashboard ReOrg - hz.dashboard --> horizon.dashboard
Rename all uses of 'hz.dashboard' in Angular code to
'horizon.dashboard'. There are two main reasons for
this change:

1. This is consistent with the pattern used for the
horizon framework (horizon.framework)

2. This aligns with the directory structure. 'horizon'
is a directory name and maps to /horizon and indicates
that the component is part of the horizon project. The
abbreviation 'hz' is now only used for legacy code.

Change-Id: Ic22d6b0819ed62b3af2656fcb167b08ac55857c2
Partial-Bug: 1478717