favicon, logo and logo-splash should allow theme overrides
favicon, logo and logo-splash are all elements that should
be easy to customize within a theme.  By adding this path
to the STATICFILES_DIR, this enables very granular overrides
of any of the images located within dashboard/img, by simply
adding an 'img' directory to your custom theme directory (or
static directory) and providing an image with the same file
name.

Closes-Bug: #1448283
Change-Id: I4feb229cac805e1b1fa3a55b5ecea1339a9fc8f9
Co-Authored-By: Tracy Jones <tjones@vmware.com>
Co-Authored-By: Diana Whitten <hurgleburgler@gmail.com>