Sidebar now inherits from a Bootstrap Theme
Horizon's sidebar now uses .nav and implements a nav using
.nav-stacked and .nav-pills to support a standard Bootstrap theme.
It was also simplfied to use Bootstrap's built in accordion
functionality and therefore removes all of the javascript required to
duplicate this experience.  There was also a small glitchy experience
with the custom accordion code, this is now solved with Bootstrap.

All highly customized style elements that were specific to the 'default'
theme were moved into the theme folder itself.  Only the styles
necessary to conform standard Bootstrap remain in the dashboard
directory tree.

The sidebar was formerly named an 'accordion-nav', but this name is
coupled with the presentational experience that is implemented within
the sidebar. Therefore, sidebar is more semantic than accordion.
Because it is now simple to override the sidebar with a non-accordion
experience, we should just call it the sidebar going forward.

sidebar-bg.png was removed as it was only limiting the ability to
easily grow the size of the sidebar.  It was providing a box-shadow
which can be done with pure css.

To test and use this functionality, clean your static directory, then
run collectstatic and compress prior to starting your server.

Partially-Implements: blueprint horizon-theme-css-reorg
Partially-Implements: blueprint bootstrap-html-standards
Change-Id: Ic7fcde7dd82e70815dc7fcae04bc9ae9d618d4d8