Add Angular Directive to show msg for empty tables
Ideally smart-table would have a default message when filtering
yields no results, but it does not.  This patch provides a general
directive that allows Angular tables to specify a custom message, or
uses a default "No items to display" message if not.

To test, set DISABLED = False in
 _3031_identity_users_panel.py, and/or
 _1051_project_ng_images_panel.py and/or
 _2081_admin_flavors_panel.py

Then go to the last Users panel in the Identity menu,
last Images Panel in Project menu, respectively.

Co-Authored-By: Rajat Vig <rajatv@thoughtworks.com>
Co-Authored-By: Cindy Lu <clu@us.ibm.com>

Change-Id: Ia086bb791d6c74af738209795195ff3ea1860981
Closes-bug: #1438822