Delete VIP action refactoring
"Delete VIP" is not a batch action so it makes sense to use
tables.Action instead of tables.DeleteAction to prevent extra
success message which is in fact incorrect. "Delete VIP" is
a specific action: there is no separate tab for VIPs. That's
why pool's id is used for this message. Moreover, a default
confirmation message is quite confusing too: "You have selected
<pool name associated with VIP>". This patch adds a custom help
text with VIP id to this message.

Closes-Bug: #1439180

Change-Id: I74c13bcbee56e22d8814e8a4118c906c23e2405e