Horizon Logo Documentation is out of date
Due to recent changes to faciliate logo and favicon overrides,
the documentation was out of date.

Also, it was noted that the theme path needed to get updated as
well because it was recently moved up a directory.

It was noted, while writing the documentation, that styles had
been placed in the 'default' theme that should be global.

Change-Id: I025514b8ff31863858a7941c2380dd8c6397772a
Closes-bug: #1485272