Babel extractor translating inner tags
The current extractor does not support nesting of inner tags.
This creates problems when we have something like this
"<translate> hello <b>beautiful</b> world</translate>".
The message extraction gets prematurely cut off.

Futhermore, the extractor does not support trimming.
This is important because angular-gettext trims the content and use that
as the msgid. The goal is to get it working with angular-gettext, hence,
we also need to trim our messages.

Change-Id: I9265f731431521240dddd24bb87baa97cc3ed4b5
Implements: blueprint babel-translate-inner-tags