Adding clone feature to Transfer Table
Currently, it is a pain to create a transfer table. A user must create two
separate templates: one for allocated and one for available. Both
templates are nearly identical with subtle differences. It would be great if
we can write one template and clone it. This would remove a lot of
boilerplate templates.

For a demo of the feature, take a look at dependent patch,
pay close attention to select-projects.html
https://review.openstack.org/#/c/202315/

Change-Id: I0c2cdae8a8994d37abaf22fa6921991c0d6485b3
Implements: blueprint transfer-table-clone-feature