Remove key-pair-details.html duplicate
In the new launch instance wizard, there appears to be an extra
key-pair-details.html under the security groups step. It is not used.

Remove:
launch-instance/security-groups/keypair-details.html

Keep:
launch-instance/keypair/keypair-details.html

Note: This was originally removed in:
https://review.openstack.org/#/c/207655/

But then was re-added during a re-org patch that didn't merge
conflict with the above.

Change-Id: Id56292c35eb8cb054a4c181ea67943fc666fb184
Closes-Bug: #1479998