Pie Charts now inherit from Bootstrap Theme
Horizon's pie charts had hard coded styles directly in the JavaScript
code.  This made it impossible for them to inherit any of their look
and feel from the install theme.  This patch fixes that problem.

It was noted, while doing the refactor, that the legend was being
implemented within D3, but this was not necessary at all, since it
has no graphing visuals associated with it.  It was impossible to
allow the legend label sizes to inherit and be configurable from
the css while they were being generated in the SVG.  The height
could never be known unless D3 knew the height of each element.

Noted Improvements:
* Consistent Colors! Colors, even for distribution charts, now come
  from the palette
* All aspects of the charts are now resizable and can inherit the
  sizes of the charts, as well as the paddings and type settings
  from the theme
* A bug in the 'Show Numbers' function was discovered and will be
  fixed as a result
* The legend symbol, which was a hardcoded cube before, is now an
  icon font, and therefore can be customized per theme.  I.E., you
  can now use rectangles, circles, ovals, smiley faces, etc
* It now inherits directly from the Theme

Change-Id: I5d5c13e57553f8b504593aac6d7611175e900f3c
Partially-Implements: blueprint horizon-theme-css-reorg