Fix usage of netaddr '.broadcast'
netaddr 0.7.16 changed the behavior of IPNetworks with /31 and /32
prefixes to make their 'broadcast' attribute return None. This patch
replaces the use of the attribute with a -1 index lookup to get the
last address instead.

Revert Ida700555c133e27da2c76357986adfc8193fe89b and re-enables the
test_routerrule_detail test.

Change-Id: I0887b0112f97b537e4c1467c47ab78b1f94f238b
Closes-Bug: 1490403
Closes-Bug: 1490423