Angular alert service - add auto dismiss
This patch auto dismisses 'info' and 'success' messages for
the Angular toast service.  It uses settings defined in
conf.js in the framework directory, which were copied over
from settings.py. We've copied over those settings
because we don't want to mix client-side settings with
other settings.  This is a first step in that goal.

To test:
1. Insert 'horizon.framework.widgets.toast.service'
as dependency (for example, into launch-instance-model.service.js)
3. Put several toastService.add(<type>, <message>);
and see that they are dismissed.
4. Also, change the auto_fade_alerts setting in conf.js
to see it adjust accordingly.

Partially Implements: blueprint launch-instance-redesign
Change-Id: I848c8e495f1a360aecc3a7353011b69aaead79e8