Reset flavors for other than "Boot from Image" source type.
If a launched instance from "Boot from image" source type is not
meeting the given image RAM/disk requirements, the "Flavor" field
of Launch Instance will display the below message.

"Some flavors not meeting minimum image requirements have been
disabled."

And the flavors that are not meeting the specified requirements
will be disabled under "Flavor" field.

After this message has been shown if we try to choose a source
type other than "Boot from image", the flavors are still in
disabled mode and message also not being cleared.

To fix this issue, Set the image name to default when the source type
is changed from "Boot from image" to other source type.So that
disableFlavorsForImage() has been called on Image name change to
reset flavors.

Change-Id: I1a105eb84bd5d92ad521d9a8ae290912d48cf275
Closes-Bug: #1482507