Topology filter out non tenant ports
This change filters out ports presented to the topology view which
are connected to networks which are not visible to the tenant.  Such
ports can be present when an admin creates an instance and ties and instance
into two networks.

Change-Id: Ib0e7ea38b42b580a65455c344f100e5aad67954e
Closes-bug: #1494358
(cherry picked from commit 42ae3dbf058c7a02dbb9d61a8e924ec1e06af3c3)