Use raw data for inline edits of boolean fields
Boolean fields use filters to turn the True/False values into Yes/No
but this makes the inline edit checkbox always checked. This patch
will use the raw data for the cell when setting the value for the
inline edit checkbox so that it is correctly checked or not checked.

Closes-Bug: #1496417
Change-Id: I9edb0a1d1b88b918ba334393c44e8389f6c785a4