Refactor status and admin state translation code
In some of the dashboard 'views.py' files, there is code that enables
the detail status and, when applicable, admin state labels to be
translated.  The current coding style of these sections is hard
to read.  Any and all sections of code that use this style should
be refactored to enhance the code readability.  This change
proposes to create a single utility method that handles this
repetitive code block that is used in many places.  Thus to perform
the translation label lookup, this method need only be called.  This
method will provide a single clean and consistent code block that can
be used for any of these situations in the future.

Change-Id: I6353abef58fc2a481d4379eb1a2f8289d7f1790b
Closes-Bug: #1490031