Made error fields in ng launch instance workflow more visible
During UX review, feedback was given that fields with errors in the new
angular launch instance workflow should have a more obvious line/glow
around them.  Currently, the "has-warning" class is being used for
fields that have errors, which gives the fields a light yellow glow
around them upon invalid input.  This change simply switches the class
applied to these fields to the "has-error" class.  This class gives the
fields a darker red glow that is easier to see and tell that an invalid
input has been entered.

Change-Id: I5a28d4429a723e68d93509a178e32fe5225bc95a
Closes-Bug: #1445729