Fixed empty "VPN Connections" value from displaying a blank value
Currently, on the VPN Service Details page, the "VPN Connections" label is
showing a blank value when no connections are found.  Displaying a blank
value here is misleading to the user and inconsistent with the "Description"
label above on that same page, which displays a "None" value for an empty
description.  This change proposes a similar solution to this issue by
displaying "None" in place of the formally blank value.

Change-Id: Ia5f26879bc6c02209028b063f36b1866d49375d2
Closes-Bug: #1500590