Fix horizon/test selenium tests for Django 1.8
This fixes the dummy user used in the Selenium tests
for Django 1.8 updates (_meta.pk.to_python() etc strike
again).

It also addresses the refactoring of horizon.utils back
into angular code. This aspect of the patch, along with
the general structure of the test base.html, needs to
be revisited, but fixing that structure correctly is
outside of the scope of this patch.

Also the ObjDictWrapper was enhanced to allow easier
debugging of test failures, and simplified to reduce
LOC achieving the same functionality.

Change-Id: I5a50ad89163a8cce06959d36314a836c4bd3592c
Closes-Bug: 1501163