Adding ifAllowed method to policy service
Policy current returns a response object containing the allowed field,
which either is true or false. To simplify this a bit, we need a wrapper
method ifAllowed that would return a promise that either resolves or
rejects. If it resolves, the success handler is called, and if it rejects, the error
handler is called.

This is more useful in practice and negates the need to check the response
object. Furthermore, this is useful for creating a hz-if-policies directive in
the future. The user table controller is included in this patch to show how it
can be use in practice.

Change-Id: I58d088cfe2df001c9252697592ad03a217cdb30a
Closes-Bug: #1491180