security group not mandatory to create instance
as per the nova api and legacy launch instance,
the security group is not mandatory option to create an instance.

but in the new launch instance form, it is marked as
mandatory and it is not allowed to create/launch
an instance if no security group was selected.

this patch making the security group is optional.

Change-Id: Ib381a16922abaf2e704c6a46a8189aec23adfbaa
Closes-Bug: #1501722