Horizon Dropdown now inherits from Bootstrap Theme
Horizon dropdowns now use proper Bootstrap markup to allow for theme
inheritance.  It was noticed during this effort that Horizon was
using a mixture of <a> and <button> elements within dropdowns, and
having to override CSS so that their styles were shared.  This
created unnecessary complexity in the CSS:

http://getbootstrap.com/components/#dropdowns

It was also noted that we were passing Bootstrap classes defined in
the Python code to the templates through a static string that was
concatenated within all the other needed attributes.  This makes
overriding the templates with custom classes difficult.  New logic was
added to allow classes to be separate.  This more granular approach
was added to the table drop down templates.

Partially-Implements: blueprint horizon-theme-css-reorg
Partially-Implements: blueprint bootstrap-html-standards
Partial-bug: #1490207

Change-Id: Id82999e5db37035968d39361ba9be4ff87c26f66