Introduce DEFAULT_SERVICE_REGIONS
It should be together with related change in django-openstack-auth, if
specified it will change the default service region calculation: it
will be taken from this setting (on a per-endpont basis) instead of a
value stored in cookies.  This value is still checked for sanity,
i.e. it should be present in Keystone service catalog.

Change-Id: I7e36f766870793f3e8fc391a06f0ee49deaa7add
Related-Bug: #1506825
Closes-Bug: #1703390