Wait until the success message popup disappears before proceeding
It seems that integration test runner became faster (forgive me
Jenkins, I was unfair to you, initially thinking that the error was
caused by your slowdown!). In times before October 7, 2015 the
auto-fading success message faded away before next actions were taken
by the test. After this date, tests started to manage (in some cases)
to click the 'Delete Xs' button that was right under the success
message before the message was gone. This effectively dismissed the
message and stole the click from button. The situation was worsened by
the fact that at the time the failure screenshot was made, a message
was already gone, leaving no evidence. Once this case was uncovered,
the solution was easy: to wait until the message disappears (combining
waiting for spinner and the message into a single function), then take
further actions.

Change-Id: I4f5743abbb3cc00852df4647fee0c68b70131833
Partial-Bug: #1505320