Improving JS unit test branch coverage
This patch is testing untested branches.

There is a bug in the Launch Instance model where it is supposed to
remove null properties from the top level of the model.  It is not doing
that properly.  So that bug is also fixed.

This patch currently raises the branch coverage in openstack_dashboard
from ~84% to ~89%.

Closes-bug: 1506891
Change-Id: I7434126246b15209b62daeff75744e0a3b435494