Add correct license for hzPasswordMatch and bi-directional check
Somewhere along the way, the license for hzPasswordMatch was
mishandled. This patch corrects  it. I am the original author of
hzPasswordMatch.

Currently, hzPasswordMatch directive does not do bi-directional check.
Meaning that the confirmation password input triggers a check on the
password input, but not vice versa. This patch makes it bi-directional.

Change-Id: Ib2907b7af7d08bb18c1f3c7daedfafccfc26c1f6
Closes-bug: #1506569
Closes-bug: #1506571